backend/requirements-dev.txt
-r requirements.txt
coverage==7.6.1
pytest==8.3.5
pytest-django==4.9.0
django-stubs==4.2.7
django-stubs-ext==4.2.7
